# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Donation Support API",
        "version": "1.0.0"
    },
    "host": "stream.local",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/mainsettings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "\u0421\u043f\u0438\u0441\u043e\u043a \u043d\u0430\u0441\u0442\u0440\u043e\u0435\u043a",
                "responses": {
                    "200": {
                        "description": "\u0423\u0441\u043f\u0435\u0448\u043d\u0430\u044f \u043e\u043f\u0435\u0440\u0430\u0446\u0438\u044f",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/MainSettings"
                            }
                        }
                    },
                    "404": {
                        "description": "\u041d\u0430\u0441\u0442\u0440\u043e\u0439\u043a\u0438 \u043d\u0435 \u043d\u0430\u0439\u0434\u0435\u043d\u044b"
                    }
                }
            }
        },
        "/api/donate/settings/{user}": {
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "\u0421\u0431\u0440\u043e\u0441 \u043d\u0430\u0441\u0442\u0440\u043e\u0435\u043a",
                "description": "\u0421\u0431\u0440\u043e\u0441 \u043d\u0430\u0441\u0442\u0440\u043e\u0435\u043a",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/MainSettings"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "User is not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "\u041f\u0440\u043e\u0444\u0438\u043b\u044c \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f",
                "description": "\u041f\u0440\u043e\u0444\u0438\u043b\u044c \u0430\u0432\u0442\u043e\u0440\u0438\u0437\u043e\u0432\u0430\u043d\u043d\u043e\u0433\u043e \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u044f",
                "responses": {
                    "200": {
                        "description": "\u0410\u043d\u043a\u0435\u0442\u043d\u044b\u0435 \u0434\u0430\u043d\u043d\u044b\u0435",
                        "schema": {
                            "$ref": "#/definitions/UserProfile"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "User is not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/fetchlinks": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "\u0421\u0441\u044b\u043b\u043a\u0438 \u0434\u043b\u044f \u043e\u0442\u043e\u0431\u0440\u0430\u0436\u0435\u043d\u0438\u044f \u0432 \u043f\u0440\u043e\u0444\u0438\u043b\u0435",
                "description": "\u0421\u043f\u0438\u0441\u043e\u043a \u0441\u0441\u044b\u043b\u043e\u043a \u043d\u0430 \u0441\u0442\u0440\u0430\u043d\u0438\u0446\u0443 \u0434\u043e\u043d\u0430\u0442\u0430, \u043f\u043e\u0434\u043f\u0438\u0441\u043a\u0443, \u0432\u0438\u0434\u0435\u043e...",
                "responses": {
                    "200": {
                        "description": "\u041c\u0430\u0441\u0441\u0438\u0432 \u0441\u0441\u044b\u043b\u043e\u043a: \u0414\u043e\u043d\u0430\u0442, \u041f\u043e\u0434\u043f\u0438\u0441\u043a\u0430, ",
                        "schema": {
                            "$ref": "#/definitions/FetchLinks"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "User is not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "Admin": {
            "properties": {
                "appends": {
                    "description": "test",
                    "type": "string",
                    "enum": [
                        "success",
                        "error"
                    ]
                }
            }
        },
        "MainSettings": {
            "properties": {
                "Donate": {
                    "properties": {
                        "commission": {
                            "type": "integer"
                        },
                        "minAmount": {
                            "type": "integer"
                        },
                        "minCommissionAmount": {
                            "type": "integer"
                        },
                        "percent": {
                            "type": "integer"
                        }
                    },
                    "type": "object"
                },
                "DonateForm": {
                    "properties": {
                        "btnBack": {
                            "properties": {
                                "bgColor": {
                                    "type": "string"
                                },
                                "flat": {
                                    "type": "boolean"
                                },
                                "fontFamily": {
                                    "type": "string"
                                },
                                "icon": {
                                    "type": "string"
                                },
                                "light": {
                                    "type": "boolean"
                                },
                                "size": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        },
                        "btnCanel": {
                            "properties": {
                                "bgColor": {
                                    "type": "string"
                                },
                                "flat": {
                                    "type": "boolean"
                                },
                                "fontFamily": {
                                    "type": "string"
                                },
                                "icon": {
                                    "type": "string"
                                },
                                "light": {
                                    "type": "boolean"
                                },
                                "size": {
                                    "type": "integer"
                                },
                                "text": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "btnDonat": {
                            "properties": {
                                "bgColor": {
                                    "type": "string"
                                },
                                "color": {
                                    "type": "string"
                                },
                                "flat": {
                                    "type": "boolean"
                                },
                                "fontFamily": {
                                    "type": "string"
                                },
                                "icon": {
                                    "type": "string"
                                },
                                "light": {
                                    "type": "boolean"
                                },
                                "size": {
                                    "type": "integer"
                                },
                                "text": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "btnNext": {
                            "properties": {
                                "bgColor": {
                                    "type": "string"
                                },
                                "color": {
                                    "type": "string"
                                },
                                "flat": {
                                    "type": "boolean"
                                },
                                "fontFamily": {
                                    "type": "string"
                                },
                                "icon": {
                                    "type": "string"
                                },
                                "light": {
                                    "type": "boolean"
                                },
                                "size": {
                                    "type": "integer"
                                },
                                "text": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "header": {
                            "properties": {
                                "bgColor": {
                                    "type": "string"
                                },
                                "color": {
                                    "type": "string"
                                },
                                "fontFamily": {
                                    "type": "string"
                                },
                                "fontSize": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        },
                        "main": {
                            "properties": {
                                "bg": {
                                    "type": "string"
                                },
                                "bgColor": {
                                    "type": "string"
                                },
                                "bgImage": {
                                    "type": "string"
                                },
                                "bgVideo": {
                                    "type": "string"
                                },
                                "justify": {
                                    "type": "string"
                                },
                                "light": {
                                    "type": "boolean"
                                },
                                "minDonate": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        },
                        "other": {
                            "properties": {
                                "donatevideolimit": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "type": "object"
                },
                "pushDonate": {
                    "properties": {
                        "amount": {
                            "properties": {
                                "color": {
                                    "type": "string"
                                },
                                "font": {
                                    "type": "string"
                                },
                                "size": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        },
                        "attachment": {
                            "properties": {
                                "src": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "image": {
                            "properties": {
                                "color": {
                                    "type": "string"
                                },
                                "size": {
                                    "type": "integer"
                                },
                                "src": {
                                    "type": "string"
                                },
                                "tile": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        },
                        "main": {
                            "properties": {
                                "bgColor": {
                                    "type": "string"
                                },
                                "duration": {
                                    "type": "integer"
                                },
                                "grid": {
                                    "properties": {
                                        "image": {
                                            "type": "string"
                                        },
                                        "xs6": {
                                            "type": "boolean"
                                        },
                                        "xs12": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                },
                                "sound": {
                                    "properties": {
                                        "src": {
                                            "type": "string"
                                        },
                                        "volume": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        },
                        "message": {
                            "properties": {
                                "color": {
                                    "type": "string"
                                },
                                "font": {
                                    "type": "string"
                                },
                                "size": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        },
                        "name": {
                            "properties": {
                                "color": {
                                    "type": "string"
                                },
                                "font": {
                                    "type": "string"
                                },
                                "size": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        },
                        "position": {
                            "properties": {
                                "attachment": {
                                    "properties": {
                                        "angle": {
                                            "type": "integer"
                                        },
                                        "classPrefix": {
                                            "type": "string"
                                        },
                                        "height": {
                                            "type": "integer"
                                        },
                                        "id": {
                                            "type": "string"
                                        },
                                        "scaleX": {
                                            "type": "integer"
                                        },
                                        "scaleY": {
                                            "type": "integer"
                                        },
                                        "width": {
                                            "type": "integer"
                                        },
                                        "x": {
                                            "type": "integer"
                                        },
                                        "y": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "counter": {
                                    "properties": {
                                        "angle": {
                                            "type": "integer"
                                        },
                                        "classPrefix": {
                                            "type": "string"
                                        },
                                        "height": {
                                            "type": "integer"
                                        },
                                        "id": {
                                            "type": "string"
                                        },
                                        "scaleX": {
                                            "type": "integer"
                                        },
                                        "scaleY": {
                                            "type": "integer"
                                        },
                                        "width": {
                                            "type": "integer"
                                        },
                                        "x": {
                                            "type": "integer"
                                        },
                                        "y": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "properties": {
                                        "angle": {
                                            "type": "integer"
                                        },
                                        "classPrefix": {
                                            "type": "string"
                                        },
                                        "height": {
                                            "type": "integer"
                                        },
                                        "id": {
                                            "type": "string"
                                        },
                                        "scaleX": {
                                            "type": "integer"
                                        },
                                        "scaleY": {
                                            "type": "integer"
                                        },
                                        "width": {
                                            "type": "integer"
                                        },
                                        "x": {
                                            "type": "integer"
                                        },
                                        "y": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "title": {
                                    "properties": {
                                        "angle": {
                                            "type": "integer"
                                        },
                                        "classPrefix": {
                                            "type": "string"
                                        },
                                        "height": {
                                            "type": "integer"
                                        },
                                        "id": {
                                            "type": "string"
                                        },
                                        "scaleX": {
                                            "type": "integer"
                                        },
                                        "scaleY": {
                                            "type": "integer"
                                        },
                                        "width": {
                                            "type": "integer"
                                        },
                                        "x": {
                                            "type": "integer"
                                        },
                                        "y": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        },
                        "video": {
                            "properties": {
                                "src": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "type": "object"
                },
                "subscribers": {
                    "properties": {
                        "costPremium": {
                            "type": "integer"
                        }
                    },
                    "type": "object"
                }
            }
        },
        "UserProfile": {
            "properties": {
                "accout_type": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "blocked_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_vk": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profile_mode": {
                    "type": "string"
                },
                "settings": {
                    "type": "object"
                },
                "social_google": {
                    "type": "boolean"
                },
                "social_twitch": {
                    "type": "boolean"
                },
                "social_vkontakte": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "stream_hash_link": {
                    "type": "string"
                },
                "twitch_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "object"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "FetchLinks": {
            "properties": {
                "[]": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "link": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "types": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                }
            }
        }
    }
}